refs #212: Avoid duplicate annotation messages
In some cases, external tools (like ghc-mod) might produce duplicate
error messages.  Updated the HaskellAnnotationHolder to track the
messages added via annotations and skip the ones which have already been
added, returning 'null' to indicate that it was not created. Note also
that this also that this makes the various 'create*Annotation' methods Nullable,
so we'll need to check the return value if passing the annotation to
something else.