fixes #204: Type info parse errors
Refactored GhcMod.parseProblems into two methods so
absolute path inference can be performed optionally.
This is mostly so we can parse errors from ghc-mod
when certain commands fail (e.g. type).

Refactored the GhcUtil.handleTypeInfo into two methods
(one now prefixed with 'unsafe') so we can throw errors
for failing to parse the type info message and retrieve
the error from ghc-mod.  The `unsafe` variant is mostly
convenient for testing.

Updated GhcMod.type and GhcModi.type methods to check for
the thrown exception from handleTypeInfo and properly report
the ghc-mod error via a tooltip or, if unable to parse the error,
report stdout as a notification to the user.