refs #407; Avoid constructing IElementType early
In IntelliJ 2019.3 the concept of 'LoadingState' has been introduced.
Checks are done in the constructor of 'IElementType' to
ensure the proper LoadingState, e.g.

    if (!(this instanceof IFileElementType)) {
      LoadingPhase.assertAtLeast(LoadingPhase.FRAME_SHOWN);
    }

This was happening because we are initializing IElementTypes in
the 'HaskellTypes' interface and then static importing
them to use them as static constants in 'HaskellLanguage'
to later be used in the 'HaskellSyntaxHighlighter'. To defer all of
this, we are now referencing the IElementTypes in the
'HaskellSyntaxHighlighter' directly.